break down top level properties.ts
