+++ src/index.ts
@@ -1,7 +1,6 @@
 // .js imports required for CommonJS
 
 export * from './primitives.js'
-export * from './properties.js'
 
 export * from './docs/index.js'
 export * from './interface/index.js'
