+++ src/interface/events.ts
@@ -1,0 +1,20 @@
+/**
+ * Interface specific events.
+ */
+ export enum InterfaceEventName {
+  APPROVE_TOKEN_TXN_SUBMITTED = 'Approve Token Transaction Submitted',
+  CONNECT_WALLET_BUTTON_CLICKED = 'Connect Wallet Button Clicked',
+  EXPLORE_BANNER_CLICKED = 'Explore Banner Clicked',
+  EXPLORE_SEARCH_SELECTED = 'Explore Search Selected',
+  EXPLORE_TOKEN_ROW_CLICKED = 'Explore Token Row Clicked',
+  NAVBAR_RESULT_SELECTED = 'Navbar Result Selected',
+  NAVBAR_SEARCH_EXITED = 'Navbar Search Exited',
+  NAVBAR_SEARCH_SELECTED = 'Navbar Search Selected',
+  TOKEN_IMPORTED = 'Token Imported',
+  TOKEN_SELECTED = 'Token Selected',
+  TOKEN_SELECTOR_OPENED = 'Token Selector Opened',
+  WALLET_CONNECT_TXN_COMPLETED = 'Wallet Connect Transaction Completed',
+  WALLET_SELECTED = 'Wallet Selected',
+  WRAP_TOKEN_TXN_INVALIDATED = 'Wrap Token Transaction Invalidated',
+  WRAP_TOKEN_TXN_SUBMITTED = 'Wrap Token Transaction Submitted',
+}
