+++ src/interface/properties.ts
@@ -1,0 +1,13 @@
+/** Known navbar search result types */
+export enum NavBarSearchTypes {
+  COLLECTION_SUGGESTION = 'collection-suggestion',
+  COLLECTION_TRENDING = 'collection-trending',
+  RECENT_SEARCH = 'recent',
+  TOKEN_SUGGESTION = 'token-suggestion',
+  TOKEN_TRENDING = 'token-trending',
+}
+
+export enum WalletConnectionResult {
+  FAILED = 'Failed',
+  SUCCEEDED = 'Succeeded',
+}
